Define scope of what these measures apply to
The question of where to draw the line around the tech we're measuring risk
for when some of it is outsourced has come up frequently in discussions with
many teams.

We can't outsource risk, even if ownership and management of technology is
outsourced, so the in-house/outsourced line doesn't represent the boundary of
our technical risk. We have a lot of custom software which is built and
maintained by suppliers for MOJ, and we're responsible for the decision to
take that approach and for the risk of that technology existing.

I hope that what's out of scope by this definition will not be a source of
technical risk for us. It may involve product risk (if
software-as-a-service doesn't meet our needs) or security risk (if it doesn't
adequately protect our data or our systems), and we're still responsible for
our decision to use this kind of technology, but we don't have responsibility
for the technology itself.

Fixes #15