Merge 4 measures about ease of making changes into 2
These are higher-level questions now than the rest of the tech measures, but
they're closer to what we were getting at with the original four. I'm not sure
yet whether it's helpful that they're a bit more general - there might be too
many factors to consider in each question for them to be answered
consistently.

Fixes #1