Add more detail to backup/recovery measure
This is more rigorous in a way which definitely applies to our larger/more
critical systems, but could well be overly onerous for others. Not sure how to
word this better without introducing more judgement at the measurement stage
though - I'd rather not ask people to decide whether the recovery strategy is
adequate for the service when answering this question. This one needs more
feedback.