Rename Atrophy group to Decay
From #14:

"Atrophy" isn't an easy term to understand, and has led to confusion multiple
times. Here are some of the risks we're trying to identify with this group:

- technology degrades over time by default
- active work and attention is needed to prevent that from happening
- factors outside your control influence how fast technology degrades
- we often know about these factors/events in advance, but can't always
  predict exactly when they will arise even if we know they're likely to
  happen in general (eg identification of security vulnerabilities,
  hardware failure)
- the rate of degradation isn't linear, and the work required to prevent it
  varies over time as well

I think "decay" is probably a better alternative heading for this group. I
don't think it captures all of those points, but it's probably more easily
understood than "atrophy".